Remove obsolete documentation for unload-plugins
unload-plugin's job was to unload plugins. This would have made use of a
hook to run the unload hooks for each plugin. However this is obsoleted
by trigger-hook, which can easily perform the same task. As such
unload-plugin is now replaced entirely by a single :unload hook as
documented in the readme.